requirements.txt
Django==1.6.1
Markdown==2.3.1
South==0.8.4
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-toolbelt==0.0.1
django-uuidfield==0.4.0
djangorestframework==2.3.12
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wheel==0.22.0
wsgiref==0.1.2
